+++ src/constants.js
@@ -210,6 +210,12 @@ export const SORT_DIRECTION = {
   DESCENDING: 'descending',
 };
 
+export const SORTABLE_SEARCH_RESULT_LIST_COLUMNS = {
+  TITLE: 'title',
+  CONTRIBUTORS: 'contributors',
+};
+
+
 export const DEFAULT_FILTERS_NUMBER = 6;
 
 export const FACETS = {