+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Added sort options in Actions for Instance/Holdings/Item. Refs UIIN-2357.
 * Statistical code empty field error. Refs UIIN-2420.
 * Hide the `Actions` button for the view page if there are no permissions. Refs UIIN-2360.
+* When duplicating an item, the circulation history is duplicated when it should not be. Fixes UIIN-2419.
 * Don't reset browse query when on Browse route and click Browse segment. Fixes UIIN-2434.
 
 ## [9.4.5](https://github.com/folio-org/ui-inventory/tree/v9.4.5) (2023-04-03)